# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI20K Agent",
    "description": "AI Agent built with LangGraph",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/admin/pipelines": {
      "get": {
        "tags": [
          "admin-pipelines"
        ],
        "summary": "List Pipelines",
        "operationId": "list_pipelines_api_v1_admin_pipelines_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelinesListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/pipelines/{dag_id}/runs": {
      "post": {
        "tags": [
          "admin-pipelines"
        ],
        "summary": "Trigger Pipeline Run",
        "description": "`dag_id` is checked here **and** inside `airflow_client` (plan's own\n\"cả route và client\") -- this route's check produces the clean `400` the\nspec wants; the client's is the chokepoint that can never be forgotten\nby a future caller. The running/queued check is a live call, never the\n10s list cache -- running a crawl pipeline twice at once is the fastest\nway to get IP-blocked by an OTA and pay for the same API calls twice.\n(A second guard against the same overlap: all 4 DAGs set\n`max_active_runs=1` in Airflow itself, so even a concurrent-POST race\nthat gets past this check can't make Airflow execute two runs at once --\nthe loser just queues behind the winner instead of running in parallel.)\n\nOnly `has_params` DAGs (Embedding, today) may pass a non-empty `conf` --\nthe other 3 have no `params` block in their DAG, so an admin-supplied key\nwould reach a task's `dag_run.conf` with no validation at all.",
        "operationId": "trigger_pipeline_run_api_v1_admin_pipelines__dag_id__runs_post",
        "parameters": [
          {
            "name": "dag_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dag Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriggerRunRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriggerRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PipelineItem": {
        "properties": {
          "dag_id": {
            "type": "string",
            "title": "Dag Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "is_paused": {
            "type": "boolean",
            "title": "Is Paused"
          },
          "schedule": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schedule"
          },
          "has_params": {
            "type": "boolean",
            "title": "Has Params"
          },
          "last_run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PipelineLastRun"
              },
              {
                "type": "null"
              }
            ]
          },
          "recent_runs": {
            "items": {
              "$ref": "#/components/schemas/PipelineRunSummary"
            },
            "type": "array",
            "title": "Recent Runs"
          }
        },
        "type": "object",
        "required": [
          "dag_id",
          "label",
          "description",
          "is_paused",
          "has_params"
        ],
        "title": "PipelineItem"
      },
      "PipelineLastRun": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "progress": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PipelineProgress"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "state",
          "start_date",
          "end_date"
        ],
        "title": "PipelineLastRun"
      },
      "PipelineProgress": {
        "properties": {
          "done": {
            "type": "integer",
            "title": "Done"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "eta_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eta Seconds"
          },
          "estimated_records": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Records"
          }
        },
        "type": "object",
        "required": [
          "done",
          "total"
        ],
        "title": "PipelineProgress"
      },
      "PipelineRunSummary": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "state"
        ],
        "title": "PipelineRunSummary"
      },
...
      "PipelinesListResponse": {
        "properties": {
          "connected": {
            "type": "boolean",
            "title": "Connected"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PipelineItem"
            },
            "type": "array",
            "title": "Items"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "const": "airflow_unavailable"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "connected",
          "items"
        ],
        "title": "PipelinesListResponse"
      },
...
      "TriggerRunRequest": {
        "properties": {
          "conf": {
            "additionalProperties": true,
            "type": "object",
            "title": "Conf"
          }
        },
        "type": "object",
        "title": "TriggerRunRequest"
      },
      "TriggerRunResponse": {
        "properties": {
          "dag_id": {
            "type": "string",
            "title": "Dag Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "state": {
            "type": "string",
            "title": "State"
          }
        },
        "type": "object",
        "required": [
          "dag_id",
          "run_id",
          "state"
        ],
        "title": "TriggerRunResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}